requirements.txt
selenium>=4.15.2
webdriver-manager>=4.0.1
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
pytest>=7.4.3
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1 